Use int instead of bigInt because that's that teamwork uses; Also rename the file so it gets run _after_ teamwork.

+++ database/migrations/2021_12_27_032000_create_event_registrations_table.php
@@ -17,7 +17,7 @@ class CreateEventRegistrationsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('event_id');
-            $table->unsignedBigInteger('team_id')->nullable();
+            $table->unsignedInteger('team_id')->nullable();
             $table->timestamps();
             $table->softDeletes();
